requirements.txt
django~=4.2.17
djangorestframework~=3.16.0